requirements.txt
libgen-api-enhanced
zlibrary
requests
aiohttp
beautifulsoup4
lxml
python-dotenv
internetarchive
diskcache
playwright
langfuse
